requirements.txt
pandas
openpyxl
python-docx
PyPDF2
pdfplumber
langchain
openai
streamlit
pytest
python-dotenv
faiss-cpu 
mlflow 
langchain-community
langchain-openai
SQLAlchemy

